fix: Program.cs のロジックを DashboardLauncher に集約しパッチカバレッジ 0% を解消
- DashboardLauncher.LaunchWithAutoDetect() を追加し、設定読み込み・
  DB パス解決・Console 出力を DashboardCommand.cs 側に集約する
- Program.cs の SetAction は LaunchWithAutoDetect() の1行呼び出しに簡略化
  （不要な using も削除）
- codecov.yml: Program.cs と DashboardCommand.cs を ignore に追加
  （外部プロセス起動コードはユニットテスト不可のため）
