+++ codecov.yml
@@ -3,10 +3,11 @@ coverage:
     patch:
       default:
         # Blazor Hybrid (WPF + Razor) UI code is excluded because UI testing is not feasible.
-        # DashboardCommand launches an external process and is not unit-testable.
+        # CLI entry point and dashboard launcher launch external processes and are not unit-testable.
         target: auto
         threshold: 5%
 
 ignore:
   - "src/CloudMigrator.Dashboard/**"
   - "src/CloudMigrator.Cli/Commands/DashboardCommand.cs"
+  - "src/CloudMigrator.Cli/Program.cs"
